[Enh] Breaking __wrapper__.py in submodule
